sentiment_analysis_system/requirements.txt
# Core dependencies
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
asyncio>=3.4.3
websockets>=10.0

# ML/NLP
transformers>=4.11.3
torch>=1.9.0
tensorflow>=2.6.0
nltk>=3.6.3
spacy>=3.1.3
scikit-learn>=1.0
textblob>=0.15.3
vaderSentiment>=3.3.2

# Data processing
pandas>=1.3.3
numpy>=1.21.2

# Database
psycopg2-binary>=2.9.1
pymongo>=3.12.0
redis>=3.5.3

# Streaming
kafka-python>=2.0.2

# API clients
tweepy>=4.4.0
praw>=7.4.0
requests>=2.26.0

# Visualization
matplotlib>=3.4.3
seaborn>=0.11.2
wordcloud>=1.8.1

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1

# Deployment
gunicorn>=20.1.0
docker>=5.0.0